% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-km.R
\docType{data}
\name{GeomKm}
\alias{GeomKm}
\alias{geom_km}
\title{Display Kaplan Meier Curve}
\format{\preformatted{Classes 'GeomKm', 'Geom', 'ggproto' <ggproto object: Class GeomKm, Geom>
    aesthetics: function
    default_aes: uneval
    draw_group: function
    draw_key: function
    draw_layer: function
    draw_panel: function
    non_missing_aes: 
    parameters: function
    required_aes: x y
    setup_data: function
    use_defaults: function
    super:  <ggproto object: Class Geom> 
}}
\usage{
GeomKm

geom_km(mapping = NULL, data = NULL, stat = "km", position = "identity",
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), is combined with the default mapping at the top level of the
plot. You only need to supply \code{mapping} if there isn't a mapping
defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. There are
  three types of arguments you can use here:

  \itemize{
  \item Aesthetics: to set an aesthetic to a fixed value, like
     \code{color = "red"} or \code{size = 3}.
  \item Other arguments to the layer, for example you override the
    default \code{stat} associated with the layer.
  \item Other arguments passed on to the stat.
  }}
}
\description{
Kaplan Meier Curve

Add a Kaplan-Meier survival curve
}
\section{Aesthetics}{

\code{geom_km} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
  \item \strong{\code{x}} The survival/censoring times. This is automatically mapped by \link{stat_km}
  \item \strong{\code{y}} The survival probability estimates. This is automatically mapped by \link{stat_km}
  smallest level in sort order is assumed to be 0, with a warning
  \item \code{alpha}
  \item \code{color}
  \item \code{linetype}
  \item \code{size}
}
}
\examples{
sex <- rbinom(250, 1, .5)
df <- data.frame(time = exp(rnorm(250, mean = sex)), status = rbinom(250, 1, .75), sex = sex)
ggplot(df, aes(x = time, status = status, color = factor(sex))) + geom_km()
}
\seealso{
The default stat for this geom is \code{\link{stat_km}} see
  that documentation for more options to control the underlying statistical transformation.
}
\keyword{datasets}

