% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkm.R
\name{ggkm}
\alias{ggkm}
\title{Creates a Kaplan-Meier plot with at risk tables below}
\usage{
ggkm(sfit, table = TRUE, returns = FALSE, xlabs = "Time-to-event",
  ylabs = "Survival (\%)", xlims = c(0, max(sfit$time)), ylims = c(0, 1),
  ystratalabs = c("CC", "CG", "GG"), ystrataname = "Genotype",
  timeby = 100, main = "", pval = FALSE, marks = TRUE, shape = 3,
  legend = TRUE, legendposition = c(0.85, 0.8), subs = NULL,
  linecols = c("red", "blue", "green"), linetype = c("solid", "dashed",
  "twodash"), ...)
}
\arguments{
\item{returns}{logical: if TRUE, return an arrangeGrob object}

\item{timeby}{numeric: control the granularity along the time-axis}

\item{main}{plot title}

\item{pval}{logical: add the pvalue to the plot?}

\item{marks}{logical: should censoring marks be added?}

\item{legend}{logical: should a legend be added to the plot?}

\item{sfit:}{a survfit object}

\item{table:}{logical: Create a table graphic below the K-M plot, indicating at-risk numbers?}

\item{xlabs:}{x-axis label}

\item{ylabs:}{y-axis label}

\item{ystratalabs:}{The strata labels. Default = levels(summary(sfit)$strata)}

\item{ystrataname:}{The legend name. Default = "Strata"}

\item{shape:}{what shape should the censoring marks be, default is a vertical line}
}
\value{
a ggplot is made. if return=TRUE, then an arrangeGlob object
is returned
}
\description{
Creates a Kaplan-Meier plot with at risk tables below
}
\examples{
 library(survival)
 data(colon)
 fit <- survfit(Surv(time,status)~rx, data=colon)
 ggkm(fit, timeby=500)
}
\author{
Abhijit Dasgupta with contributions by Gil Tomas & Michael Way
\url{http://statbandit.wordpress.com/2011/03/08/an-enhanced-kaplan-meier-plot/}
slight adjustment to cope with none strata calls (e.g. Surv(time,event)~1),
option to remove the legend and also draw marks at censoring locations by Nadieh Bremer
}

